pyproject.toml
[tool.poetry]
name = "data_fetchers"
version = "0.1.0"
description = "Data Fetchers"
authors = ["qmlai <quantitativeml@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "data_fetchers" }]

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.3.2"
pandas = "^2.3.1"
python-dotenv = "^1.1.1"
entsoe-py = "^0.7.4"
pyarrow = "^21.0.0"
lxml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"

[build-system]
requires = ["poetry-core>=1.5.1"]
build-backend = "poetry.core.masonry.api"